% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdn.R
\name{calendar2jdn}
\alias{calendar2jdn}
\alias{jdn2calendar}
\title{jdn}
\usage{
calendar2jdn(date, jdn = c("jdn", "mjd"), calendar = c("gregorian", "julian"))

jdn2calendar(
  n,
  jdn = c("jdn", "mjd"),
  calendar = c("gregorian", "julian"),
  sep = "/"
)
}
\arguments{
\item{date}{Date to be converted. It should always include the year, month, and day.}

\item{jdn}{Select Julius Date Number or Modified Julius Date Number.}

\item{calendar}{Select whether the calendar is Gregorian or Julian.}

\item{n}{Julian Date Number or Modified Julian Date Number value.}

\item{sep}{The way the date is separated when the last date is written.}
}
\description{
Functions for calculating Julian Date from Gregorian or Julian Date Number and functions for calculating Gregorian or Julian Date from Julian Date Number.
}
\references{
Fliegel, H. F. and Van Flandern, T. C., "A Machine Algorithm for Processing Calendar Dates," Communications of the ACM 11, p. 657, 1968.

Hatcher, D. A., Simple formulae for Julian day numbers and calendar dates, Quarterly Journal of the Royal Astronomical Society, v. 25, p. 53-55, 1984.

Meeus, J., Astronomical Algorithms, 1998.
}
