% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wareki.R
\name{seireki2kyureki}
\alias{seireki2kyureki}
\alias{kyureki2seireki}
\title{kyureki}
\usage{
seireki2kyureki(
  date,
  calendar = c("gregorian", "julian"),
  era = c("south", "north", "east", "heishi", "kyoto", "non"),
  one = c("kanji", "number"),
  leap.month = c("kanji", "hiragana"),
  sep = c("kanji", "/", "-")
)

kyureki2seireki(
  date,
  calendar = c("gregorian", "julian"),
  era = c("south", "north", "east", "heishi", "kyoto", "non"),
  sep = "/"
)
}
\arguments{
\item{date}{Date to be converted. It should always include the year, month, and day.}

\item{calendar}{Select whether the Western calendar is Gregorian or Julian.}

\item{era}{Whether the Gengo of the Japanese calendar is the Southern or Northern dynasty system, the system used in the Kanto region, the Heike, and Kyoto respectively, or no era names.}

\item{one}{Whether to write the year in kanji as the first year or leave it as a number.}

\item{leap.month}{Whether to write leap months in Kanji or Hiragana.}

\item{sep}{The way the date is separated when the last date is written.}
}
\value{
String type date.
}
\description{
This is a set of functions for mutual conversion between Kyureki that the lunar calendar and the western calendar, the former calendar system of Japan. Kyureki the lunar calendar here is a calendar system that was never actually used, assuming that the Tenpo calendar was extended after the 6th year of Meiji.
}
\references{
The data for the Kyureki calendar and era names refer to the following data.

\url{https://github.com/manakai/data-locale}
}
